Updated location and sitemap
Updated location and sitemap

+++ database/migrations/2025_09_08_074508_remove_contact_person_from_advertisers_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('advertisers', function (Blueprint $table) {
-            $table->dropColumn('contact_person');
-        });
+        if (Schema::hasColumn('advertisers', 'contact_person')) {
+            Schema::table('advertisers', function (Blueprint $table) {
+                $table->dropColumn('contact_person');
+            });
+        }
     }
 
     /**
